proxyclient/spmi: support gen 4
